requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
prompt-toolkit==1.0.15
pymessenger==0.0.7.0
python-firebase==1.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wit==5.1.0
